Include a clone of the FileSystemProxy project
Before this, the FileSystemProxy files (and their built libraries) were
included as files commited to this repo. For now, switch over to using
this project as a clone in a sub-directory. This allows for simultaneous
development of the two projects.

After looking into `git submodule` and `git subtree`, I couldn't find an
advantage over simply including a local clone directly in the project.