Collapse two wildcomp() functions into one
The prior code had two versions of wildcomp(), one case-sensitive and
one case-insensitive. This change eliminates the explicit
case-insensitive version. For case-insensitive comparisons, both the
pattern and the string should be lowercased prior to comparison.

This is a better solution, because case-insensitive comparison is no
longer trivial in the face of Unicode. Appropriate case-insensitive
comparison will be dealt with later.

In the event that character-by-character comparison is desired, we might
choose to take an addition function argument in the future that performs
character conversion. I'm not yet certain how case sensitivity plays out
for all possible languages.