Broken, but compiling
Coming back to this project, it's a mess of conflicting design ideas.

UTF-8
One problem is that Windows likes to use wchar_t for Unicode support,
while UTF-8 char types are used on POSIX systems. I partially started a
migration to UTF-8 in this code, but I'm not even sure that all works.

FileSystemProxy
Another issue with the current state is that it's trying to use the
FileSystemProxy library. The purpose of the FileSystemProxy library (my
code) is to support three different file systems: Windows,
POSIX/Linux/Mac, and a mock file system for testing and debugging. This
is all changing now with the introduction of the new C++17 `filesystem`
library.

For future development, this is the clear direction. Now that this code
is compile clean, I can begin incremental migration.