Move getGroomedPattern to impl helper from private
This function used to be a private method, and set a bunch of private
member variables. This is still an ugly form, but not it's a pure helper
function, and ready to convert over to wstring from wchar_t buffers.

It leaves some ugly dynamic buffer cleanup in the PathMatcher::match()
method, but that will also be cleaned up later.